fix(nyz): fix pendulum act_scale and add random_collect_size args support(ci skip)

+++ dizoo/classic_control/pendulum/envs/pendulum_env.py
@@ -60,7 +60,7 @@ class PendulumEnv(BaseEnv):
         obs, rew, done, info = self._env.step(action)
         self._final_eval_reward += rew
         obs = to_ndarray(obs).astype(np.float32)
-        rew = to_ndarray([rew])  # wrapped to be transfered to a array with shape (1,)
+        rew = to_ndarray([rew]).astype(np.float32)  # wrapped to be transfered to a array with shape (1,)
         if done:
             info['final_eval_reward'] = self._final_eval_reward
         return BaseEnvTimestep(obs, rew, done, info)
@@ -71,7 +71,7 @@ class PendulumEnv(BaseEnv):
         self._replay_path = replay_path
 
     def random_action(self) -> np.ndarray:
-        return self.action_space.sample()
+        return self.action_space.sample().astype(np.float32)
 
     @property
     def observation_space(self) -> gym.spaces.Space:
